test: migrate main test runner to TypeScript
